\name{melt_var}
\alias{melt_var}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Melt matrix to dataframe, for IO only
}
\description{
Melt total output matrices to dataframes, identified by exporter ("from.country"), importer ("to.country"), and export sector ("from.sector").
}
\usage{
melt_var(matrix.in)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{matrix.in}{
}
}
\details{
"matrix.in" is a total output matrix (not a direct requirements/total requirements matrix)
}
