\name{multiply_FD}
\alias{multiply_FD}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Multiply IO matrix by FD vector to get total output
}
\description{
Multiply technical coefficients (i.e. direct requirements or total requirements) by FD vector (total demand)
\usage{
get_total(in.df)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{in.df}{
}
}
\details{
"in.df" is a dataframe that is the output of melt_var
}
